requirements.txt
Flask==0.11.1
django-dotenv==1.4.1 # Read environment variables from .env
django-getenv==1.3.1 # Get environment variables
newrelic
flask-sitemap==0.2.0

# Error tracking
blinker==1.4
rollbar==0.13.6

# Testing
flake8==3.0.4 # Python linting
codeclimate-test-reporter==0.1.2 # Test coverage
